Separate ElementVirtualLinked from its controller
